UIFC-473-integration-typescript fix errors after rebase
